third_party/blink/web_tests/fast/spatial-navigation/snav-unit-overflow-and-scroll-in-direction-expected.txt

Start
End
PASS successfullyParsed is true

TEST COMPLETE

PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS true is true
PASS gFocusedDocument.activeElement.getAttribute("id") is "end"
